Test for nodata tile incorrect in FetchBlock (caching MRF)
Marked as TODO: the test for no data in fetchblock.  The way it is
implemented currently only works if the NDV is the same for all bands in
the band interleaved case.
This is not a fatal bug, just not optimal, it will likely compress and
write pages for interleaved data when the NDV value is different.  It
will fail to write a when the first bad NDV is the content of all the
bands.