Added JPEG-Zen extension
JPEG tiles will include the Zen app chunk, which ensures that the
locations with value of zero across all bands do not change.  This is
accomplished by storing a 2D bitmask, RLE compressed, inside the JPEG.
A Zen app chunk with a length of zero means that the tile contains no
full zeros.